fix some bugs

+++ src/app/components/AttendanceWidget.jsx
@@ -23,7 +23,7 @@ const AttendanceWidget = () => {
 
   // Format date as needed
   const formatDate = (date) => {
-    const options = { year: 'numeric', month: 'short', day: 'numeric' };
+    const options = { year: 'numeric', month: 'long', day: 'numeric' };
     return date.toLocaleDateString(undefined, options);
   };
